# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Castle Escape API",
    "description": "Authoritative game server for Castle Escape (KTU T120B516). Lobby and inspection over REST, real-time play over the SignalR hub at /hubs/game. Endpoints that act as a player need the X-Player-Token header returned by create/join.",
    "version": "v1"
  },
  "paths": {
...
    "/api/sessions/{sessionId}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Get a session",
        "description": "Phase, players, chosen characters and current level.",
        "operationId": "GetSession",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sessions/{sessionId}/input": {
      "post": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Set the held direction",
        "description": "Same as the hub\u0027s SetDirection. Send a direction on key down and None on key up. Applied at the start of the next tick.",
        "operationId": "SubmitDirection",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "PlayerToken": []
          }
        ]
      }
    },
    "/api/sessions/{sessionId}/restart": {
      "post": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Restart the current level",
        "description": "D8: the level returns to its generated layout; both players\u0027 lives and score return to the level start.",
        "operationId": "RestartLevel",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "PlayerToken": []
          }
        ]
      }
    },
    "/api/sessions/{sessionId}/state": {
      "get": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Latest game state",
        "description": "The latest per-tick state: players, zombies, items, levers, door. JSON by default; \u0060?format=xml\u0060 or \u0060Accept: application/xml\u0060 for XML (NET-2, Adapter pattern).",
        "operationId": "GetGameState",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickStateMessage"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TickStateMessage"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/messages": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "summary": "Read buffered messages (polling channel)",
        "description": "The same messages SignalR pushes, kept in a per-session buffer for clients without a hub connection (Bridge pattern: the polling channel). Pass the last \u0060seq\u0060 you saw as \u0060afterSeq\u0060. Each \u0060body\u0060 is serialized as \u0060format=json|xml\u0060. StateUpdated is buffered every \u0060Realtime:PollingStateEveryNthTick\u0060 ticks; events are never skipped.",
        "operationId": "PollMessages",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "afterSeq",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int64"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolledMessagesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/level": {
      "get": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Static layout of the current level",
        "description": "Rows of map legend characters, plus the legend itself.",
        "operationId": "GetLevelLayout",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LevelLayoutResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/hud": {
      "get": {
        "tags": [
          "Gameplay"
        ],
        "summary": "View game state (HUD table)",
        "description": "Per player: lives, score, active powers with time left, combos, statistics. Plus level, levers and door status.",
        "operationId": "GetHud",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HudResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Direction": {
        "enum": [
          "None",
          "Up",
          "Down",
          "Left",
          "Right"
        ],
        "description": "A grid direction. \u0060None\u0060 means \u0022not moving\u0022 (key released)."
      },
      "DirectionRequest": {
        "required": [
          "direction"
        ],
        "type": "object",
        "properties": {
          "direction": {
            "description": "Up/Down/Left/Right on key down; None on key up.",
            "$ref": "#/components/schemas/Direction"
          }
        },
        "description": "Body of \u0060POST /api/sessions/{id}/input\u0060: the direction key now held."
      },
...
      "HudResponse": {
        "required": [
          "sessionId",
          "phase",
          "levelIndex",
          "maxLevel",
          "tick",
          "levelElapsedSeconds",
          "leversActive",
          "leversTotal",
          "doorOpen",
          "itemsRemaining",
          "players"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Session id.",
            "format": "uuid"
          },
          "phase": {
            "description": "Current phase.",
            "$ref": "#/components/schemas/SessionPhase"
          },
          "levelIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Current level.",
            "format": "int32"
          },
          "maxLevel": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Levels in the run.",
            "format": "int32"
          },
          "tick": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Server tick.",
            "format": "int64"
          },
          "levelElapsedSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?(?:[eE][\u002B-]?\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Time spent on the current level.",
            "format": "double"
          },
          "leversActive": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Levers active right now.",
            "format": "int32"
          },
          "leversTotal": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Levers in the level.",
            "format": "int32"
          },
          "doorOpen": {
            "type": "boolean",
            "description": "Whether the exit door is open."
          },
          "itemsRemaining": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Items still on the map.",
            "format": "int32"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HudPlayerDto"
            },
            "description": "One row per player."
          }
        },
        "description": "The \u0022View game state\u0022 table (view-game-state activity diagram)."
      },
...
      "LevelLayoutResponse": {
        "required": [
          "sessionId",
          "levelIndex",
          "theme",
          "seed",
          "width",
          "height",
          "rows",
          "legend",
          "obstacles"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Session id.",
            "format": "uuid"
          },
          "levelIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "1-10.",
            "format": "int32"
          },
          "theme": {
            "description": "Dungeon or Crypt.",
            "$ref": "#/components/schemas/LevelTheme"
          },
          "seed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Generation seed (same seed, same layout).",
            "format": "int32"
          },
          "width": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Width in tiles.",
            "format": "int32"
          },
          "height": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Height in tiles.",
            "format": "int32"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "One string per row, map legend characters."
          },
          "legend": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Meaning of every legend character."
          },
          "obstacles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LevelObstacleDto"
            },
            "description": "The themed obstacle variants this level uses (one family per theme)."
          }
        },
        "description": "Static layout of the current level (\u0060GET /api/sessions/{id}/level\u0060)."
      },
...
      "PlayerStateDto": {
        "required": [
          "playerId",
          "slot",
          "name",
          "characterId",
          "tileX",
          "tileY",
          "x",
          "y",
          "facing",
          "isMoving",
          "lives",
          "maxLives",
          "score",
          "powers",
          "combos"
        ],
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "format": "uuid"
          },
          "slot": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "characterId": {
            "type": "string"
          },
          "tileX": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tileY": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "x": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?(?:[eE][\u002B-]?\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "y": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?(?:[eE][\u002B-]?\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "facing": {
            "$ref": "#/components/schemas/Direction"
          },
          "isMoving": {
            "type": "boolean"
          },
          "lives": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "maxLives": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "score": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "powers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivePowerDto"
            }
          },
          "combos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PowerType"
            }
          }
        },
        "description": "A player in the tick state. \u0060X\u0060/\u0060Y\u0060 are interpolated for smooth rendering."
      },
...
      "PolledMessagesResponse": {
        "required": [
          "sessionId",
          "contentType",
          "latestSeq",
          "messages"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Session id.",
            "format": "uuid"
          },
          "contentType": {
            "type": "string",
            "description": "Format of every \u0060Body\u0060: application/json or application/xml."
          },
          "latestSeq": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Highest sequence number returned (pass it as \u0060afterSeq\u0060 next time); the request\u0027s \u0060afterSeq\u0060 if none.",
            "format": "int64"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolledMessageDto"
            },
            "description": "The buffered messages after \u0060afterSeq\u0060, oldest first."
          }
        },
        "description": "Messages read from the polling channel (\u0060GET /api/sessions/{id}/messages\u0060)."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SessionDto": {
        "required": [
          "sessionId",
          "joinCode",
          "phase",
          "levelIndex",
          "maxLevel",
          "players"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Session id.",
            "format": "uuid"
          },
          "joinCode": {
            "type": "string",
            "description": "Code for the second player."
          },
          "phase": {
            "description": "Current lifecycle phase.",
            "$ref": "#/components/schemas/SessionPhase"
          },
          "levelIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Current level (1-10), or 0 before the first level.",
            "format": "int32"
          },
          "maxLevel": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Number of levels in the run.",
            "format": "int32"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionPlayerDto"
            },
            "description": "Joined players in slot order."
          }
        },
        "description": "Lobby view of a session."
      },
      "SessionPhase": {
        "enum": [
          "WaitingForPlayers",
          "CharacterSelect",
          "LoadingLevel",
          "Playing",
          "LevelComplete",
          "Victory",
          "Defeat",
          "Aborted"
        ],
        "description": "Lifecycle of a two-player session."
      },
      "SessionPlayerDto": {
        "required": [
          "playerId",
          "slot",
          "name",
          "characterId",
          "connected"
        ],
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "description": "Player id.",
            "format": "uuid"
          },
          "slot": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "1 or 2; decides the start tile.",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "characterId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Chosen character, or null while choosing."
          },
          "connected": {
            "type": "boolean",
            "description": "Whether the player has a live hub connection."
          }
        },
        "description": "A player as seen in the lobby."
      },
...
      "TickStateMessage": {
        "required": [
          "sessionId",
          "seq",
          "tick",
          "phase",
          "levelIndex",
          "doorOpen",
          "players",
          "zombies",
          "items",
          "levers"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Session id.",
            "format": "uuid"
          },
          "seq": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Message sequence number.",
            "format": "int64"
          },
          "tick": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Server tick.",
            "format": "int64"
          },
          "phase": {
            "description": "Current phase.",
            "$ref": "#/components/schemas/SessionPhase"
          },
          "levelIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Current level.",
            "format": "int32"
          },
          "doorOpen": {
            "type": "boolean",
            "description": "Whether the exit door is open."
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerStateDto"
            },
            "description": "Both players."
          },
          "zombies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZombieStateDto"
            },
            "description": "All zombies."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemStateDto"
            },
            "description": "Items still on the map."
          },
          "levers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeverStateDto"
            },
            "description": "All levers."
          }
        },
        "description": "Dynamic state, sent every tick while playing."
      },
...
      "ZombieStateDto": {
        "required": [
          "id",
          "typeId",
          "tileX",
          "tileY",
          "x",
          "y",
          "facing",
          "strategy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "typeId": {
            "type": "string"
          },
          "tileX": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tileY": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "x": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?(?:[eE][\u002B-]?\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "y": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?(?:[eE][\u002B-]?\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "facing": {
            "$ref": "#/components/schemas/Direction"
          },
          "strategy": {
            "$ref": "#/components/schemas/MovementStrategyKind"
          }
        },
        "description": "A zombie in the tick state. \u0060Strategy\u0060 is how it chases right now (it can be swapped at runtime)."
      },
...
    },
    "securitySchemes": {
      "PlayerToken": {
        "type": "apiKey",
        "description": "The playerToken returned by POST /api/sessions or POST /api/sessions/join.",
        "name": "X-Player-Token",
        "in": "header"
      }
    }
  },
...
}